fix: a whole bunch of issues
