+++ src/components/overlay/overlay.test.tsx
@@ -1,0 +1,7 @@
+import { Overlay } from '.'
+
+describe('Overlay', () => {
+  it('is truthy', () => {
+    expect(Overlay).toBeTruthy()
+  })
+})
