+++ src/components/text/index.test.tsx
@@ -1,0 +1,7 @@
+import { Text } from '.'
+
+describe('Text', () => {
+  it('is truthy', () => {
+    expect(Text).toBeTruthy()
+  })
+})
