+++ src/index.tsx
@@ -4,9 +4,10 @@ export { Checkbox, Radio, RadioGroup } from './components/checkbox'
 export { Container } from './components/container'
 export { Button } from './components/button'
 export { Dropdown, DropdownItem } from './components/dropdown'
-export { Overlay } from './components/overlay'
+export * from './components/overlay'
 export * from './components/portal'
 export { Tooltip } from './components/tooltip'
+export { Text } from './components/text'
 export {
   Arrow,
   ArrowToBottom,
